Minor cleanup and refactoring
* models and tuskeys blades now show how many models and keys there are total, respectively
* moved static functions from FilePath into LaravelFileManager
* FileManageController - some refactoring
* Fixed some bugs with how some orphaned data is detected

+++ src/Http/Controllers/FileManageController.php
@@ -44,9 +44,10 @@ class FileManageController extends BaseController {
      * @param Response $response
      * @return void
      */
-    protected function responseOrView(Request $request, Response $response, string $view = 'laravel-filemanager::files'){
+    protected function responseOrView(Response $response, string $view = 'laravel-filemanager::files'){
 
         // send JSON:
+        $request = request();
         if(!$request->acceptsHtml()) return $response;
 
         // send HTML view:
@@ -54,6 +55,7 @@ class FileManageController extends BaseController {
         $vals['baseUrl'] = app('laravel-filemanager')->baseUrl($request);
         return view($view, $vals);
     }
+
     /**
      * If API request, a JSON response is sent. Otherwise, the user
      * is redirected to a blade view.
@@ -62,15 +64,27 @@ class FileManageController extends BaseController {
      * @param Response $response
      * @return void
      */
-    protected function responseOrRedirect(Request $request, Response $response, string $overridePath = ""){
+    protected function responseOrRedirect(Response $response, string $overridePath = '/'){
         
+        $request = request();
+
         // send JSON:
         if(!$request->acceptsHtml()) return $response;
         
         // redirect to HTML view:
-        $path = $overridePath ? $overridePath : (new FilePath($request))->getDir();
         $prefix = config('laravel-filemanager.head.prefix', '/file-manager');
-        return redirect("{$prefix}/browse?path={$path}");
+        
+        $ref_from_head = $request->headers->get('referer');
+        
+        $referrer = preg_replace('/^[^\?]+(\/[a-zA-Z0-9\-_]+\??)(.*)?$/','$1', $ref_from_head);
+
+        $endpoint = $referrer ?? '/';
+
+        //echo $endpoint ."  from " . $ref_from_head; exit;
+
+        $path = $overridePath ?? (new FilePath($request))->getDir();
+
+        return redirect("{$prefix}{$endpoint}?path={$path}");
     }
 
     // shows files and folders within a directory
@@ -81,7 +95,7 @@ class FileManageController extends BaseController {
             return $this->repo->downloadFile($path);
         }
         $response = $this->repo->listItemsInDir($path);
-        return $this->responseOrView($request, $response);
+        return $this->responseOrView($response,'laravel-filemanager::files');
     }
     public function models(Request $request){
         
@@ -91,12 +105,13 @@ class FileManageController extends BaseController {
             if(!$file->file_exists()) $orphaned []= $file->toArray();
         }
         $response = response([
+            'total_models' => $all->count(),
             'orphaned_models' => $orphaned,
             'path' => '/',
             'baseUrl' => app('laravel-filemanager')->baseUrl($request),
         ], 200);
 
-        return $this->responseOrView($request, $response,'laravel-filemanager::models');
+        return $this->responseOrView($response,'laravel-filemanager::models');
     }
     // adds a file to the database, only meta-data
     public function add(Request $request){
@@ -104,7 +119,7 @@ class FileManageController extends BaseController {
         
         $response = $this->repo->addModel($path);
 
-        return $this->responseOrRedirect($request, $response);
+        return $this->responseOrRedirect($response);
     }
     // removes a file from the database, does not delete the file
     public function remove(Request $request){
@@ -115,13 +130,14 @@ class FileManageController extends BaseController {
         $response = $this->repo->removeModel($file);
 
         $path = $file ? $file->dir_path : '/';
-        return $this->responseOrRedirect($request, $response, $path);
+        return $this->responseOrRedirect($response, $path);
     }
     public function delete(Request $request){
         $path = new FilePath($request);
         
+
         $response = $this->repo->deleteFile($path);
-        return $this->responseOrRedirect($request, $response, $path->getDir());
+        return $this->responseOrRedirect($response, $path->getDir());
     }
     public function newdir(Request $request){
         
@@ -131,7 +147,7 @@ class FileManageController extends BaseController {
         ]);
         $response = $this->repo->makeDir($path, $vals['name']);
         
-        return $this->responseOrRedirect($request, $response, $path->getPathRelative());
+        return $this->responseOrRedirect($response, $path->getPathRelative());
     }
     // renames a resource
     // $path->rename() is called
@@ -147,13 +163,21 @@ class FileManageController extends BaseController {
         ]);
         $response = $this->repo->rename($path, $vals['name']);
 
-        return $this->responseOrRedirect($request, $response, $path->getDir());
+        return $this->responseOrRedirect($response, $path->getDir());
     }
     public function tusUploads(Request $request){
+
+        $lfm = app('laravel-filemanager');
+
+        $keys_found = 0;
+
+        $keys_found = count(app('tus-server')->getCache()->keys());
+
         return view('laravel-filemanager::tuskeys', [
-            'baseUrl' => app('laravel-filemanager')->baseUrl($request),
+            'baseUrl' => $lfm->baseUrl($request),
             'path' => '/',
-            'orphaned_tuskeys' => FilePath::getOrphanedTusKeys(),
+            'orphaned_tuskeys' => $lfm->getOrphanedTusKeys(),
+            'total_keys' => $keys_found,
         ]);
     }
     public function tusUpload(Request $request){
@@ -176,13 +200,19 @@ class FileManageController extends BaseController {
         // find key in cache
         $cached_file = $cache->get($id, true);
 
-        // delete the key
-        $isDeleted = $cache->delete($id);
+        if(config('laravel-filemanager.debug.disable_cleanup', false)){
+            
+            // pretend to delete the key
+            $isDeleted = true;
+        } else {
+            // delete the key for real
+            $isDeleted = $cache->delete($id);
+        }
 
         $response = $isDeleted ? response([], 200) : response('Key not found', HttpResponse::HTTP_GONE);
-        $path = $isDeleted ? FilePath::relative_path($cached_file['file_path']) : '/';
+        $path = $isDeleted ? app('laravel-filemanager')->relative_path($cached_file['file_path']) : '/';
 
         // redirect
-        return $this->responseOrRedirect($request, $response, $path);
+        return $this->responseOrRedirect($response, $path);
     }
 }
